refactor: extract sysinfo module, add debug logging to silent catches (PY.2)
- New sysinfo.py (245 lines): all system info gathering extracted from
  client.py (CPU benchmark, OS detection, memory, disk, uptime)
- client.py reduced from 1,285 to 1,064 lines
- 10 silent exception catches now emit logger.debug with exc_info
- 167 tests pass, mypy clean
